+++ packages/carbon-graphs/CHANGELOG.md
@@ -4,13 +4,12 @@
 
 * Changed
   * For consistency replaced `document` in all draw helpers with `graph.legendSVG` for better control.
-
-* Changed
   * For consistency updated reflow in Graph and Gantt constructs to update the eventlines.
   * Added code to handle null/undefined/blank in paired result graph during both initial load and reflow.
   * Updated `packages/carbon-graphs/README.md` so that it has more essential information about Carbon.
   * Added array handling for loadContent on Construct based graphs.
-
+  * Throw an error when null/undefined/blank is passed as y value.
+  
 ## 2.15.0 - (November 24, 2020)
 
 * Changed